Add files via upload
Elements of the rendered content of a React component can be bound against properties of the component’s state. Whenever we update the values of state, we automatically update the content that references its attributes. I was able to update the rendered content of a component dynamically from data fetched from a third-party API.

The process is as follows:

Establish an object data within the component’s state
Establish the content to be rendered with sections referencing attributes in data
Create an event handler that fetches data from an API
On success of the API call, update the component’s state
Used jQuery as our library to make an HTTP request against an API endpoint.

For this example, you can fetch the latest Bitcoin price from CoinDesk’s public API. The content will include the price of USD, GBP, and EUR against BTC (Bitcoin). The JSON structure of the data returned by the API endpoint

#webDevelopment #fronEndWebDevelopment #clientSideFramework #React #Bitcoin #DeFi #DecentralizedFianance